# internal/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "nft-forward Public API",
    "version": "1.0.0",
    "description": "Token 化的稳定契约面。所有请求带 `Authorization: Bearer <token>`（openapi.json 除外）。\n成功返回 {\"data\":...}，失败返回 {\"error\":{\"code\",\"message\"}}。\ntoken 有 read/readwrite 两种 scope；写端点需 readwrite。\ntoken 还绑定用户角色：user 角色只能访问 /my/* 自助端点；admin 角色可访问集群观测与用户管理端点。\n"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "unauthorized",
                  "forbidden",
                  "scope_required",
                  "rate_limited",
                  "validation",
                  "not_found",
                  "conflict",
                  "duplicate_chain_node",
                  "internal"
                ]
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/info": {
      "get": {
        "summary": "当前 token 归属用户的自述信息",
        "description": "任意角色、read scope 即可。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "用户自述信息"
          },
          "401": {
            "description": "未鉴权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/probe": {
      "get": {
        "summary": "单跳连通性探测",
        "description": "任意角色、read scope 即可。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "target",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "node",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "proto",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "探测结果"
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/probe-chain": {
      "get": {
        "summary": "按规则整链逐跳探测",
        "description": "任意角色、read scope 即可。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "rule",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "逐跳探测结果"
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/my/nodes": {
      "get": {
        "summary": "我被授权使用的节点",
        "description": "user 角色、read scope。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "节点列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Node"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "非 user 角色",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/my/rules": {
      "get": {
        "summary": "我的规则列表",
        "description": "user 角色、read scope。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "规则列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Rule"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "新建我的规则",
        "description": "user 角色、readwrite scope。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "幂等键；相同键的重试回放首次创建的规则，不重复分配端口。"
          },
          {
            "name": "dry_run",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "1"
              ]
            },
            "description": "与 body.dry_run 等效，只预览不落地。"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleWriteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "新建的规则（或 dry_run 预览）"
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "端口/链路冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/my/rules/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer"
          }
        }
      ],
      "get": {
        "summary": "我的单条规则详情",
        "description": "user 角色、read scope。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "规则详情"
          },
          "404": {
            "description": "不存在或不属于本人",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "更新我的规则",
        "description": "user 角色、readwrite scope。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleWriteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新后的规则"
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "不存在或不属于本人",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "删除我的规则",
        "description": "user 角色、readwrite scope。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "删除结果"
          },
          "404": {
            "description": "不存在或不属于本人",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/nodes": {
      "get": {
        "summary": "全部节点（含未授权给当前用户的）",
        "description": "admin 角色、read scope。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "节点列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Node"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "非 admin 角色",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/dashboard": {
      "get": {
        "summary": "集群概览",
        "description": "admin 角色、read scope。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "概览统计"
          },
          "403": {
            "description": "非 admin 角色",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/usage": {
      "get": {
        "summary": "一次性计费快照",
        "description": "admin 角色、read scope。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "按用户/按节点的用量汇总与集群总量"
          },
          "403": {
            "description": "非 admin 角色",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/landing-usage": {
      "get": {
        "summary": "全用户落地出口用量台账快照",
        "description": "admin 角色、read scope 即可。每行恰含 user_id/host/port/protocol/present/used_bytes/updated_at。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "台账行数组"
          }
        }
      }
    },
...
  }
}